+++ resources-ktx/src/main/kotlin/Text.kt
@@ -3,9 +3,11 @@
 package com.redmadrobot.extensions.resources
 
 import android.content.Context
+import android.os.Parcelable
 import android.view.View
 import androidx.annotation.StringRes
 import androidx.fragment.app.Fragment
+import kotlinx.parcelize.Parcelize
 
 /**
  * Wrapper to make it possible to work with plain [String] and [StringRes] in the same way.
@@ -19,7 +21,7 @@ import androidx.fragment.app.Fragment
  *  val messageText = getString(message)
  * ```
  */
-public sealed class Text {
+public sealed class Text : Parcelable {
 
     /** Retrieves [String] using given [context]. */
     public abstract fun get(context: Context): String
@@ -28,11 +30,13 @@ public sealed class Text {
     abstract override fun hashCode(): Int
 
     /** Plain string. */
+    @Parcelize
     public data class Plain(public val string: String) : Text() {
         override fun get(context: Context): String = string
     }
 
     /** String resource, requires [Context] to get [String]. */
+    @Parcelize
     public data class Resource(@StringRes public val resourceId: Int) : Text() {
         override fun get(context: Context): String = context.getString(resourceId)
     }