+++ resources-ktx/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 
 - **Text**: added abstract methods `equals`, `hashCode`
+- **Text**: added `Parcelable` implementation
 
 ## [1.3.1-0] (2021-10-03)